graduate llama4 to core (#1865)
This PR also
- updates the README of `deepseek_v3` folder.
- move the `generate_permute_indices` triton kernel to
`torchtitan/models/moe` so that core doesn't depend on `experiments`
- deprecate the gpu checkpoint conversion scripts as now we natively
support loading checkpoint from HF using GPUs (although it is only using
GPU when doing online conversion right before training starts)
