+++ torchtitan/models/__init__.py
@@ -4,4 +4,4 @@
 # This source code is licensed under the BSD-style license found in the
 # LICENSE file in the root directory of this source tree.
 
-_supported_models = frozenset(["llama3", "llama3_ft", "deepseek_v3", "qwen3"])
+_supported_models = frozenset(["deepseek_v3", "llama3", "llama3_ft", "llama4", "qwen3"])
