+++ torchtitan/models/moe/__init__.py
@@ -4,6 +4,6 @@
 # This source code is licensed under the BSD-style license found in the
 # LICENSE file in the root directory of this source tree.
 
-_supported_experiments = frozenset(
-    ["flux", "simple_fsdp.llama3", "simple_fsdp.deepseek_v3", "vlm"]
-)
+from .moe import FeedForward, MoE, MoEArgs
+
+__all__ = ["FeedForward", "MoE", "MoEArgs"]
